Added Judges.html/UnfinishedPage.html, Adjusted Zoom
Added 2 html pages (both temporary): 
Judges.html is a simple page with nested lists for Judge names and positions.
UnfinishedPage.html is just a html with an image to be linked to pages that haven't been completed yet (Ex. Schedule.html, PresidentsLetter.html).
Adjusted the Android Zoom to a maximum of 4x.